Fix a batch of real compiler warnings from CI build logs
- vectorOfStruct_c.cpp.in / vectorOfObject_c.cpp.in: the generated
  VectorOf*GetSize() always returned std::vector::size() (size_t)
  from a function declared to return int, triggering MSVC C4267 on
  every one of the several dozen VectorOf* native wrappers. Same
  narrowing on the PushMulti call in the struct template. Added
  explicit (int) casts; these vectors never approach INT_MAX
  elements, so this is a no-op at runtime.
- bioinspired_c.cpp: cveRetinaCreate passed its `double` reduction/
  sampling parameters straight into cv::bioinspired::Retina::create,
  whose signature takes float. Added explicit casts.
- tesseract.cpp: cveTessBaseAPISetImage passed cv::Mat::elemSize()/
  step (both size_t) into Tesseract::SetImage's int parameters.
  Added explicit casts.
- ButtonTextImagePage.cs: replaced the obsolete
  Page.DisplayActionSheet with DisplayActionSheetAsync.
- AutoTestVideoIO.cs: replaced two inexact FileStream.Read calls
  (CA2022 - not guaranteed to fill the buffer) with ReadExactly.

Verified: cvextern rebuilds clean on Linux with the regenerated
vector_*.cpp files and the two direct source edits; the C# test
project and the Maui UI project's netstandard2.0 TFM build clean
with 0 warnings for the touched files.

Left out for now: the ~200 C4275 warnings from OpenCV's own
exception.hpp (third-party header, not ours to fix), the .NET
trimming/AOT analysis warnings (IL2026/IL2075/IL2091/IL2062) in
Emgu.Util/Emgu.CV reflection code (real but requires either
suppression attributes or a larger refactor), NuGet TFM-support and
MSB3270 architecture-mismatch warnings (require version/config
decisions), and the size_t/int property-type mismatches in a couple
of *.g.cpp property getters (would need touching the generic
CREATE_OCV_CLASS_PROPERTY macro shared by every generated property
across the codebase - higher risk, needs its own pass).

+++ Emgu.CV.Extern/bioinspired/bioinspired_c.cpp
@@ -12,7 +12,7 @@ cv::bioinspired::Retina* cveRetinaCreate(cv::Size* inputSize, const bool colorMo
 	try
 	{
 	#ifdef HAVE_OPENCV_BIOINSPIRED
-		cv::Ptr<cv::bioinspired::Retina> ptr = cv::bioinspired::Retina::create(*inputSize, colorMode, colorSamplingMethod, useRetinaLogSampling, reductionFactor, samplingStrength);
+		cv::Ptr<cv::bioinspired::Retina> ptr = cv::bioinspired::Retina::create(*inputSize, colorMode, colorSamplingMethod, useRetinaLogSampling, (float) reductionFactor, (float) samplingStrength);
 		*sharedPtr = new cv::Ptr<cv::bioinspired::Retina>(ptr);
 		return ptr.get();
 	#else
